Now works for all cases
Fixed for loop in generateBreakpoints. Loop only needs to iterate for
the number of interior areas, rather than all breakpoints, since the end
pins are taken care of outside of the loop; 1 line above and below for
first and last ideal break point, respectively. This removes the
necessity for finding the number of breakpoints in the first place
(whose algorithm was, admittedly, incorrect as well).